+++ modules/Material/AwesomeIcon.qml
@@ -24,6 +24,7 @@ Item {
 
     property string name
     property bool rotate: widget.name.match(/.*-rotate/) !== null
+    property bool valid: text.implicitWidth > 0
 
     property alias color: text.color
     property int size: units.dp(24)