+++ CHANGELOG.md
@@ -7,6 +7,7 @@ QML Material CHANGELOG
  * [Tabs] Allow switching tabs programmatically in a page
  * [Demo] Organize the different demos into tabs
  * [Tabs] Implement the MD spec for coloring tab text
+ * [Icon] Add a valid property to Icon and AwesomeIcon
 
 ### Version 0.0.6 (April 16, 2015)